Tweaks and docs
